Move SubmitBlockPayloadInfo into ChaCha20Wrapper as a nested struct
