--- src/LLP/include/submit_block_payload_info.hpp
@@ -1,7 +1,0 @@
-#ifndef NEXUSMINER_LLP_SUBMIT_BLOCK_PAYLOAD_INFO_HPP
-#define NEXUSMINER_LLP_SUBMIT_BLOCK_PAYLOAD_INFO_HPP
-
-// Forwarding header — canonical definition is in protocol/submit_block_payload_info.hpp
-#include "protocol/submit_block_payload_info.hpp"
-
-#endif // NEXUSMINER_LLP_SUBMIT_BLOCK_PAYLOAD_INFO_HPP
