+++ src/protocol/inc/protocol/chacha20_wrapper.hpp
@@ -3,10 +3,10 @@
 
 #include <vector>
 #include <cstdint>
+#include <cstddef>
 #include <string>
 #include <memory>
 #include "spdlog/spdlog.h"
-#include "protocol/submit_block_payload_info.hpp"
 
 namespace nexusminer {
 namespace protocol {
@@ -23,7 +23,40 @@ namespace protocol {
  */
 class ChaCha20Wrapper {
 public:
-    
+
+    /**
+     * @brief Channel-aware payload metadata for SUBMIT_BLOCK sizing.
+     *
+     * Computes expected plaintext and encrypted sizes from real inputs rather
+     * than assuming a universal fixed Tritium payload size.
+     *
+     * Hash:   plaintext = base_block_size + timestamp_size + sig_len_field_size + signature_size
+     * Prime:  plaintext = base_block_size + offset_bytes_count + timestamp_size + sig_len_field_size + signature_size
+     * encrypted = plaintext + CHACHA20_OVERHEAD (nonce 12 + tag 16 = 28)
+     */
+    struct SubmitBlockPayloadInfo {
+        uint32_t channel{0};              ///< 1 = Prime, 2 = Hash
+        size_t   base_block_size{0};      ///< 216 for Tritium (empty block body)
+        size_t   offset_bytes_count{0};   ///< 0 for Hash, variable for Prime (vOffsets.size())
+        size_t   timestamp_size{8};       ///< always 8 (uint64_t LE)
+        size_t   sig_len_field_size{2};   ///< always 2 (uint16_t LE)
+        size_t   signature_size{0};       ///< parsed/actual Falcon signature length
+
+        /// ChaCha20-Poly1305 overhead: nonce(12) + auth_tag(16)
+        static constexpr size_t CHACHA20_OVERHEAD = 28;
+
+        /// Expected plaintext size: block + offsets + timestamp + sig_len + sig
+        size_t expected_plaintext_size() const {
+            return base_block_size + offset_bytes_count
+                 + timestamp_size + sig_len_field_size + signature_size;
+        }
+
+        /// Expected encrypted size: plaintext + ChaCha20 overhead
+        size_t expected_encrypted_size() const {
+            return expected_plaintext_size() + CHACHA20_OVERHEAD;
+        }
+    };
+
     /**
      * @brief Encryption/Decryption result structure
      */
